pyproject.toml
[tool.poetry]
name = "notes-yandex-check-fastapi"
version = "0.1.0"
description = ""
authors = ["Kettariec <kettariec@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.115.3"}
sqlalchemy = "^2.0.36"
alembic = "^1.13.3"
psycopg2-binary = "^2.9.10"
asyncpg = "^0.30.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
jwt = "^1.3.1"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
